+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.2.7-beta.0
 
 - Pass correct working directory to SwiftLint #24
 - Kill remaining SwiftLint processes on deactivate #24